requirements.txt
opencv-python>=4.5.0
mediapipe>=0.8.0
scikit-learn>=1.0.0
numpy>=1.20.0
pyttsx3>=2.90
Pillow>=8.0.0
pyinstaller>=4.0  # Optional: for building executable